# Perf: Workspace pre-sizing for varlen and selection pack