## Summary
- Adds workspace pre-sizing for varlen and selection pack operations to reduce memory reallocations
- Configurable via environment variables for different workload requirements
- Behavior-preserving optimization with significant performance improvements

## Changes
- `_get_varlen_workspace()`: Honors `NSA_VARLEN_RESERVE_{N,K}` environment variables
- `grouped_selection_attention_packed()`: Honors `NSA_SEL_PACK_RESERVE_{N,L}` environment variables
- Pre-allocates workspace buffers based on expected maximum sizes
- Eliminates reallocation overhead during long-context processing

## Performance Impact
Decode benchmark results (A100 80GB):
| Context | With Pre-sizing | Baseline | Improvement |
|---------|----------------|----------|-------------|
| 512     | 5.93 ms        | 17.18 ms | -65.5%      |
| 1024    | 6.05 ms        | 15.67 ms | -61.4%      |
| 2048    | 6.41 ms        | 30.50 ms | -79.0%      |

## Test Results
✅ All tests pass on CUDA:
- Selection v2 equivalence: 52/52 tests pass
- Sliding NaN CUDA: 7/7 tests pass  
- Core invariants: 6/6 tests pass
- FA-2 parity: 3/3 tests pass

## Configuration
Optional environment variables for tuning:
- `NSA_SEL_PACK_RESERVE_N`: Pre-reserve rows for selection packing
- `NSA_SEL_PACK_RESERVE_L`: Pre-reserve max row length
- `NSA_VARLEN_RESERVE_N`: Pre-reserve total Q rows
- `NSA_VARLEN_RESERVE_K`: Pre-reserve total packed K/V size

When not set, behavior is identical to current implementation.

## Validation Report
Full GPU validation report available at: `Documentation/Reports/2025-08-29 Test Engineer Report - Workspace Pre-sizing GPU Validation v1.md`

🤖 Generated with [Claude Code](https://claude.ai/code)